content/canvas/test/test_size.attributes.style.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes.style</title>
<!-- Testing: Canvas size is independent of CSS resizing -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="50" height="30" style="width: 100px; height: 50px"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.width == 50, "canvas.width == 50");
ok(canvas.height == 30, "canvas.height == 30");

SimpleTest.finish();

});
</script>
